rename file to be able to git add the file was called Application.js and every time I saved it it would let me add so i couldn't commit. Terminal history: $ git add Application.js The following paths are ignored by one of your .gitignore files: Application.js Use -f if you really want to add them. ~/wdi/projects/tic-tac-toe (feature) $ git add Main.js ~/wdi/projects/tic-tac-toe (feature)
I finally renamed it and it worked. Don't understand this .gitignore
response.

This file first included only comments of the break down of the problem.
I then began to fill parts with code.
The addSymbol function concept exists twice. Did not realize I was doing
this until I saw the other. No grunt nag has been done on any code as it
wasn't considered part of the git.

Next commit will be the aftermath of running the grunt nag.